feat(editor): decompose EditorPane/Toolbar/BubbleMenu monoliths (v1-05)
Extract 8 modules from 3 monolith files:
- editorPaneHelpers.ts: constants, types, pure utility functions
- useEntityCompletion.ts: entity completion hook with candidate resolution
- useEditorKeybindings.ts: hotkey registrations and slash command routing
- useEditorSetup.ts: TipTap init, store wiring, autosave, content loading
- InlineAiOverlay.tsx: inline AI overlay component + request helpers
- ToolbarFormatGroup.tsx: toolbar button/separator/icons/item definitions
- BubbleMenuFormatActions.tsx: format buttons + link input UI
- BubbleMenuAiActions.tsx: AI skill buttons

Line counts after decomposition:
- EditorPane.tsx: 1515 → 226 lines
- EditorToolbar.tsx: 457 → 97 lines
- EditorBubbleMenu.tsx: 414 → 170 lines

All re-exports preserved for backward compatibility.
22 decomposition guard tests (AC-1 through AC-8).

Closes #1179
